Menu fixed, rotation fixed, STOPPED fixed
The menu was not selectable in the previous version; that has now
been fixed.

When the device was rotated, the graphs would lose their state.
They no longer lose their state.

On some devices, STOPPED appeared on the displays of the
altimeter and VSI even though they were clearly receiving data.
Fixed.

The Altimeter and VSI classes have now been split to follow the
MVP pattern. Altimeter and VerticalSpeedIndicator now contain
the business logic (they are effectively the "Model"). MainActivity
implements the Presenter functions and AltimeterView and
VSIView implement the views.

Changed in accordance with the code analysis performed by
Android Studio.